added container for dialogs components
